重构操作记录。

+++ common-lib/src/main/java/com/smartindustry/common/mapper/sm/StorageRecordMapper.java
@@ -27,5 +27,5 @@ public interface StorageRecordMapper extends BaseMapper<StorageRecordPO, Long> {
      * @param sid
      * @return
      */
-    List<StorageRecordPO> queryBySid(Byte sid);
+    List<StorageRecordPO> queryBySid(Long sid);
 }
